Split code into single responsibility modules
